+++ GoTorrent/include/torrent_struct.h
@@ -1,5 +1,5 @@
 struct TorrentStruct{
-    int length;
+    long long length;
     int piece_length;
     char* announce;
     int announce_length;
@@ -7,7 +7,7 @@ struct TorrentStruct{
     char* name;
     int name_length;
     char* pieceSHA;  
-    int piece_count;
+    long long piece_count;
 };
 
 struct TrackRespStruct{
